backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-pptx>=0.6.23
Pillow>=10.2.0
lxml>=4.9.3
python-multipart>=0.0.6
pydantic>=2.10.0

